requirements.txt
django
djangorestframework
django_cleanup
gtts
networkx
matplotlib
django-cors-headers
sympy